Refactor and add redux dependencies.
