+++ cadg-ui/src/components/header-content/header-content.js
