+++ cadg-ui/src/components/main-content/main-content.js
